feat(google-analytics): Added google analytics data pipeline
Refactors common functionality out.
