+++ data_pipeline/crossref_event_data/helper_class.py
@@ -1,4 +1,6 @@
-from data_pipeline.utils import update_deployment_env_placeholder
+from data_pipeline.utils.pipeline_config import (
+    update_deployment_env_placeholder
+)
 
 
 # pylint: disable=too-few-public-methods, too-many-instance-attributes
