+++ data_pipeline/spreadsheet_data/google_spreadsheet_config.py
@@ -1,5 +1,7 @@
 from data_pipeline.utils.csv.config import BaseCsvConfig
-from data_pipeline.utils import update_deployment_env_placeholder
+from data_pipeline.utils.pipeline_config import (
+    update_deployment_env_placeholder
+)
 
 
 class MultiSpreadsheetConfig:
