+++ data_pipeline/spreadsheet_data/google_spreadsheet_etl.py
@@ -1,7 +1,5 @@
-import json
 import re
 from tempfile import NamedTemporaryFile
-from typing import Iterable
 
 from google.cloud.bigquery import WriteDisposition
 
@@ -23,6 +21,7 @@ from data_pipeline.utils.csv.metadata_schema import (
 from data_pipeline.utils.data_pipeline_timestamp import (
     get_current_timestamp_as_string
 )
+from data_pipeline.utils.pipeline_file_io import write_jsonl_to_file
 
 
 def etl_google_spreadsheet(spreadsheet_config: MultiCsvSheet):
@@ -154,7 +153,7 @@ def transform_load_data(
         record_metadata,
         standardized_csv_header,
     )
-    write_to_file(processed_record, full_temp_file_location)
+    write_jsonl_to_file(processed_record, full_temp_file_location)
     write_disposition = get_write_disposition(csv_sheet_config)
     load_file_into_bq(
         filename=full_temp_file_location,
@@ -164,13 +163,6 @@ def transform_load_data(
         write_mode=write_disposition,
         project_name=csv_sheet_config.gcp_project,
     )
-
-
-def write_to_file(json_list: Iterable, full_temp_file_location: str):
-    with open(full_temp_file_location, "w") as write_file:
-        for record in json_list:
-            write_file.write(json.dumps(record, ensure_ascii=False))
-            write_file.write("\n")
 
 
 def standardize_field_name(field_name: str):
